backend/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
appdirs==1.4.4
Authlib==1.3.0
bcrypt==5.0.0
beautifulsoup4==4.14.3
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.3
curl_cffi==0.13.0
ecdsa==0.19.1
exceptiongroup==1.3.1
fastapi==0.109.0
frozendict==2.4.7
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httptools==0.7.1
httpx==0.27.0
idna==3.11
itsdangerous==2.1.2
lxml==6.0.2
multitasking==0.0.12
numpy==1.26.3
pandas==2.1.4
passlib==1.7.4
peewee==3.19.0
platformdirs==4.5.1
protobuf==6.33.3
psycopg2-binary==2.9.11
pyasn1==0.6.1
pycparser==2.23
pydantic==2.5.3
pydantic_core==2.14.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.1
SQLAlchemy==2.0.25
starlette==0.35.1
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.27.0
uvloop==0.22.1
watchfiles==1.1.1
webencodings==0.5.1
websockets==16.0
yfinance==1.0
